Move dependencies to devDependencies
There were some things in seneca which were put in dependencies instead
of devDependencie for testing purposes. Since the tests seem to fail
already anyways, there is little point in keeping it in an
inappropriate location.

There were vulnerabilities with some of these packages, since they
were in the wrong section in the `package.json` you'd get an audit
error with `snyk`.